feature: add service interfaces and its implementation for Payment entity

+++ src/main/java/com/portfolio/lmf_api/dto/RequestPaymentDTO.java
@@ -11,5 +11,5 @@ import lombok.Setter;
 @Setter
 public class RequestPaymentDTO {
     private Integer amount;
-    private Integer matchId;
+    private Long matchId;
 }
